+++ app/routes/components/CodeHighlight/changelog.codegen.ts
@@ -9,4 +9,6 @@ export const changelog = ''
   + '- 2025-10-16: Refactor code to use hast to build the HTML. That make the\n'
   + '  `CodeHighlight` component support `lineHighlight` prop, etc.\n'
   + '- 2025-10-22: Refactor code. And use JSS for styling.\n'
+  + '- 2025-12-08: Support the highlighting for CSS language, remove the dependency\n'
+  + '  on `ThemeProvider`, etc.\n'
   ;