+++ app/routes/components/CodeHighlight/index.tsx
@@ -1,9 +1,10 @@
 import { useState } from "react";
+
 import { DocLayout } from "@/layouts/DocLayout/doc-layout";
+
 import type { Route } from "./+types/index";
 import { changelog } from "./changelog.codegen";
-import Demo01 from "./demos/Demo01";
-import demo01SourceCode from "./demos/Demo01.source_code.codegen";
+import Doc from "./doc.mdx";
 import { sourceCode } from "./source_code.codegen";
 
 export function meta(_: Route.MetaArgs) {
@@ -36,11 +37,11 @@ export default function CodeHighlightPage() {
       />
       {tabsValue === "doc" && (
         <DocLayout.Content>
-          <DocLayout.Live node={<Demo01 />} code={demo01SourceCode} />
+          <Doc />
         </DocLayout.Content>
       )}
       {tabsValue === "source" && (
-        <DocLayout.Content>
+        <DocLayout.Content variant="files">
           <DocLayout.Files files={sourceCode} />
         </DocLayout.Content>
       )}